fastapi_app/requirements.txt
sqlmodel==0.0.11
fastapi==0.104.1
pydantic==1.10.13
uvicorn==0.24.0.post1
PyJWT[crypto]==2.8.0
requests== 2.31.0
JPype1==1.4.1
SPARQLWrapper== 2.0.0
itsdangerous==2.1.2
werkzeug==3.0.1
cachetools==5.3.2


